Adding in the `vac.` element.
I'm adding in this element to the translation syntax to allow for
vacat's to be recorded in the translation. We make some use of these
in AIO translations, e.g. where tribe or deme names are indented.

These `vac.` will only be transformed at the beginning of a line at
present, but as this is sufficient for our use case, i'm not going to
look into it any further at present.