Modify vac. and the way <lb/> are generated.
I've changed vac. to take a character value and have a low uncertainty
as it seems that this will generally fit our use case after discussion
with Pietro. Other alternatives either do not give us enough information
to carry out multiple levels of indentation, or are too far from
appropriate syntax.

I've also changed how <lb/> are generated, so that text on a newline
with at least two spaces will generate the <lb/> This means that we're
not having to introduce a lot of extra blank lines to the translations,
and we can maintain an indentation scheme that makes the Leiden+ text
more readable and easier to distinguish different sections.